refactor: package구조 변경
